+++ html/40x.html
@@ -1,1 +1,1 @@
-404 not found
+400 something has gone wrong!!
