+++ html/bad_request.html
@@ -1,0 +1,1 @@
+bad request!
