+++ src/cgi/cgi.cpp
@@ -1,25 +1,3 @@
 #include "cgi.hpp"
 
-
-// Cgi::Cgi(char **env){
-//     for (int i = 0; env[i] != nullptr; ++i) {
-//         env_vector.push_back(env[i]);
-//     }
-// }
-
-// Cgi::~Cgi(){}
-
-// pid_t Cgi::execute(Cgi& cgi)
-// {
-//     cgi.pid = fork();
-
-//     if(cgi.pid == -1)
-//     {
-//         return -1;
-//     }
-//     else if (cgi.pid == 0)
-//     {
-//         // dup2()
-//     }
-// }
-// //임시파일 ?  미니쉘 랜덤파일 ?? 
+void
