+++ src/tcp/AScket.cpp
@@ -9,8 +9,13 @@ ASocket::ASocket( void ) {
 	addr.sin_addr.s_addr = INADDR_ANY;
 	addr.sin_port = htons( 8080 );
 
+
 	if ( bind( sock, ( const struct sockaddr* )&addr, sizeof( addr ) ) == ERROR )
 		throw err_t( "fail to bind" );
+
+	// int optval = 1;
+    // setsockopt(sock, SOL_SOCKET, SO_REUSEADDR, &optval,
+    //            sizeof(optval));
 }
 
 ASocket::ASocket( socket_t sockServer ) {
